refactor: #183 Removed duplicates & sorted components
